+++ front-end/src/MyFridge.js
@@ -1,73 +1,27 @@
 import React, { useState } from "react";
 import "./MyFridge.css";
 import { groups } from "./fridgeDataFake";
-import { dot, chipDays, chipAmount } from "./itemColoring";
-import Modal from "./deleteModal";
+// import { dot, chipDays, chipAmount } from "./itemColoring";
+// import Modal from "./deleteModal";
+import MyFridgeItem from "./MyFridgeItem";
+// import { countItems } from "./countItems";
 
 const MyFridge = (props) => {
-  const [show, setShow] = useState(false);
-  const [itemName, setItemName] = useState("Hello");
-  const [itemId, setItemId] = useState(0);
+  // const [show, setShow] = useState(false);
+  // const [itemName, setItemName] = useState("Hello");
+  // const [itemId, setItemId] = useState(0);
+  // const [itemCount, setItemCount] = useState(countItems(groups));
 
-  const renderItem = (data, j) => {
-    const rowEvent = (event) => {
-      const title = event.currentTarget.getAttribute("title");
-      const id = event.currentTarget.getAttribute("id");
-      // https://stackoverflow.com/questions/43335452/pass-item-data-to-a-react-modal
-      setItemName(title);
-      setItemId(id);
-      setShow(true);
-    };
-
-    const onDelete = (itemInput, groups) => {
-      let matchIndex = parseInt(itemId);
-      for (let i = 0; i < groups.length; i++) {
-        var removeIndex = groups[i].object
-          .map(function (item) {
-            return item.id;
-          })
-          .indexOf(matchIndex);
-        if (removeIndex !== -1) {
-          groups[i].object.splice(removeIndex, 1);
-        }
-      }
-    };
-
+  const renderItem = (data) => {
     // dot, chipAmount, and chipDays can be found in itemColoring.js
-    return (
-      <tbody key={j}>
-        <tr>
-          <td>
-            <span>{dot(data.daysleft)}</span>
-            <span className="title">{data.title}</span>
-            <span>{chipAmount(data.amount, data.daysleft)}</span>
-            <span>{chipDays(data.daysleft)}</span>
-          </td>
-          <td>
-            <button
-              title={data.title}
-              id={data.id}
-              amount={data.amount}
-              daysleft={data.daysleft}
-              onClick={rowEvent}
-            >
-              x
-            </button>
-          </td>
-        </tr>
-        <Modal
-          onClose={() => setShow(false)}
-          show={show}
-          onDelete={() => onDelete(data, groups)}
-          itemName={itemName}
-        />
-      </tbody>
-    );
+    return <MyFridgeItem key={data.id} food={data} />;
   };
 
   // groups is an object in fridgeDataFake.js
   return (
     <div>
+      {/* <p>You Have {itemCount} Items in your Fridge</p> */}
+      {/* <countItems groups={groups} /> */}
       {groups.map((item, i) => (
         <div key={i}>
           <h2 className="header">{item.header}</h2>
