+++ front-end/src/countItems.js
@@ -1,0 +1,15 @@
+// import React, { useState } from "react";
+
+// export function countItems(groups) {
+//   let count = 0;
+//   for (let i = 0; i < groups.length; i++) {
+//     alert("YO" + i);
+//     for (let j = 0; j < groups[i].object; j++) {
+//       count = count + 1;
+//     }
+//     alert("YPP" + count);
+//     return count;
+//   }
+// }
+
+// export default countItems;
